requirements.txt
PyJWT==1.3.1
bcrypt==4.2.1
passlib==1.7.4
uvicorn==0.34.0
fastapi==0.115.6
pydantic==2.10.5
sqlalchemy==2.0.37
python-dotenv==1.0.1